+++ src/main/java/org/xdef/impl/compile/CompileXScript.java
@@ -112,7 +112,7 @@ final class CompileXScript extends CompileStatement {
 	 * @param sc XData model of attribute, text, comment, PI.
 	 * @param parent parent model
 	 */
-	final void genTemplateData(final XCodeDescriptor sc, final XNode parent) {
+	final void genTemplateData(final XData sc, final XNode parent) {
 		if (parent != null && parent.getKind() == XMNode.XMELEMENT) {
 			XElement p = (XElement) parent;
 			sc.copyOptions(p);
@@ -155,6 +155,7 @@ final class CompileXScript extends CompileStatement {
 			_g.internalMethod("eq", 1);
 			_g.topToBool();
 			_g.genStop();
+			sc.setValueType(XD_STRING, "eq");
 			sc._onAbsence = sc._onFalse = _g._lastCodeIndex + 1;
 			_g.genLDC(defTmp);
 			if ("$text".equals(sc.getName())) {
@@ -190,7 +191,6 @@ final class CompileXScript extends CompileStatement {
 			_g.genStop();
 			_g._sp  = -1;
 			_g._mode = gmode;
-			sc.setValueType(XD_STRING, "string");
 		} else {
 			compileCheckExpression(sc);
 		}
@@ -205,7 +205,8 @@ final class CompileXScript extends CompileStatement {
 		_g._mode = CompileBase.TEXT_MODE;
 		sc.clearOptions();
 		sc.clearActions();
-		XOccurrence occ = new XOccurrence();
+		sc.setValueType(XD_STRING, "string");
+		XOccurrence occ = new XOccurrence(); // undefined occurrence
 		while (_sym != NOCHAR) {
 			if (_sym == SEMICOLON_SYM) {
 				nextSymbol();
@@ -261,7 +262,6 @@ final class CompileXScript extends CompileStatement {
 								&& _g._code.get(addr + 3).getCode() == STOP_OP){
 								XDValue value = _g._code.get(addr + 1);
 								XDParser p = (XDParser) _g._code.get(check);
-								sc.setValueType(p.parsedType(), p.parserName());
 								XDParseResult r =
 									p.check(null,value.toString());
 								if (r.errors()) {
@@ -270,8 +270,6 @@ final class CompileXScript extends CompileStatement {
 							} else {
 								sc.setValueType(XD_STRING, "string");
 							}
-						} else {
-							sc.setValueType(XD_STRING, "string");
 						}
 						sc._check = _g._lastCodeIndex + 1;
 						_g.addCode(new CodeI1(XD_STRING, CALL_OP, addr), 1);
@@ -384,15 +382,14 @@ final class CompileXScript extends CompileStatement {
 			errorAndSkip(XDEF.XDEF425, SCRIPT_SEPARATORS); //Script error
 		}
 		sc.setOccurrence(occ);
-		if (occ.isFixed() && sc._deflt >= 0) {
+		if (!sc.isSpecified()) {
+			sc.setOptional();
+		} else if (sc.isFixed() && sc._deflt >= 0) {
 			//Both actions 'fixed' and 'default' can't be specified
 			error(XDEF.XDEF415);
 			return;
 		}
-		if (!sc.isSpecified() && sc._deflt >= 0) {
-			sc.setMinOccur(0);
-		}
-		if (sc._onIllegalAttr != -1 && !occ.isIllegal() && sc._match < 0) {
+		if (sc._onIllegalAttr != -1 && !sc.isIllegal() && sc._match < 0) {
 			//Not allowed script for '&{0}'
 			warning(XDEF.XDEF494, "onIllegalAttr");
 		}
@@ -418,7 +415,6 @@ final class CompileXScript extends CompileStatement {
 			error(XDEF.XDEF422); //Duplicated script section
 		}
 		sc._check = -2;
-		sc.setValueType(XD_STRING, "string"); //default type
 		if (_sym == SEMICOLON_SYM) { //';'
 			nextSymbol();
 			return;
@@ -434,9 +430,7 @@ final class CompileXScript extends CompileStatement {
 			sc._check = check;
 			//we try to set type of checked object and the type method name.
 			sc.setValueType(CompileBase.getParsedType(typeName), typeName);
-			if (_g.getVariable(typeName) != null) {
-				sc.setRefTypeName(typeName);
-			}
+			XDParser p = null;
 			if (_g._lastCodeIndex > 0 && check < _g._code.size()) {
 				XDValue y = _g._code.get(check);
 				if (y.getCode() == CALL_OP) {// execute type method
@@ -448,11 +442,24 @@ final class CompileXScript extends CompileStatement {
 							&& y.getItemId() == XD_PARSER) {
 							j = _g._code.get(i+2).getCode();
 							if (j == PARSERESULT_MATCH || j == STOP_OP) {
-								XDParser p = (XDParser) y;
-								sc.setValueType(p.parsedType(), p.parserName());
+								p = (XDParser) y;
+//								sc.setValueType(p.parsedType(), p.parserName());
 							}
 						}
 					}
+				} else if (y instanceof XDParser) {
+					p = (XDParser) y;
+				}
+				if (p != null) {
+					sc.setValueType(p.parsedType(), p.toString());
+				}
+				if (_g.getVariable(typeName)==null && typeName.indexOf('.')>0) {
+					String s = typeName.substring(0, typeName.lastIndexOf('.'));
+					if (_g.getVariable(s) != null) {
+						sc.setRefTypeName(s);
+					}
+				} else if (typeName != null && _g.getVariable(typeName) != null) {
+					sc.setRefTypeName(typeName);
 				}
 			}
 		}
@@ -1225,6 +1232,13 @@ final class CompileXScript extends CompileStatement {
 					&& (result.getName().equals("$text")
 					|| result.getName().equals("$textcontent"))) {
 					result._cdata = 'T';
+					if ("textcontent".equals(result.getName())
+						&& result.maxOccurs() > 1) {
+						//Maximum occurrence in "xd:textcontent" attribute
+						// can not be higher then 1
+						warning(XDEF.XDEF535);
+						result.setOccurrence(result.minOccurs(), 1);
+					}
 				} else {
 					//The token '&{0}' is not allowed here
 					error(XDEF.XDEF411, _idName);