+++ src/test/java/buildtools/GenXComponents.java
@@ -1,0 +1,13 @@
+package buildtools;
+
+import test.xdef.TestXComponentsGen;
+
+/** Generate XComponents for tests.
+ * @author Vaclav Trojan
+ */
+public class GenXComponents {
+	/** @param args not used. */
+	public static void main(String... args) {
+		TestXComponentsGen.main();
+	}
+}