+++ src/test/java/test/xdef/component/XCa.java
@@ -204,7 +204,7 @@ public static class s implements org.xdef.component.XComponent{
     XD_XPos=x.getXPos();
     XD_Model=x.getXMElement().getXDPosition();
     XD_Object = (XD_Parent=p)!=null ? p.xGetObject() : null;
-    if (!"F13967C40588CF06660803F245A14449".equals(
+    if (!"5FCE23176BEF462CFADEB6B692814384".equals(
       x.getXMElement().getDigest())) { //incompatible element model
       throw new org.xdef.sys.SRuntimeException(
         org.xdef.msg.XDEF.XDEF374);
@@ -326,7 +326,7 @@ public static class s_1 implements org.xdef.component.XComponent{
     XD_XPos=x.getXPos();
     XD_Model=x.getXMElement().getXDPosition();
     XD_Object = (XD_Parent=p)!=null ? p.xGetObject() : null;
-    if (!"F13967C40588CF06660803F245A14449".equals(
+    if (!"5FCE23176BEF462CFADEB6B692814384".equals(
       x.getXMElement().getDigest())) { //incompatible element model
       throw new org.xdef.sys.SRuntimeException(
         org.xdef.msg.XDEF.XDEF374);